requirements.txt
python-telegram-bot~=20.0
yookassa
requests
pytz
python-telegram-bot[job-queue]==20.0
python-dotenv